+++ common/mouse_reltemposlider.c
@@ -21,7 +21,7 @@ static void update_statusbar(struct Tracker_Windows *window){
 
 static int MoveRelTempoSlider_Mouse(
 	struct Tracker_Windows *window,
-	int x,int y
+	float x,float y
 ){
 	struct WBlocks *wblock=window->wblock;
